Make the cancellation flag a size_t, not a uint32_t

+++ lib/src/atomic.h
@@ -7,7 +7,7 @@
 
 #include <windows.h>
 
-static inline uint32_t atomic_load(const volatile uint32_t *p) {
+static inline size_t atomic_load(const volatile size_t *p) {
   return *p;
 }
 
@@ -21,7 +21,7 @@ static inline uint32_t atomic_dec(volatile uint32_t *p) {
 
 #else
 
-static inline uint32_t atomic_load(const volatile uint32_t *p) {
+static inline size_t atomic_load(const volatile size_t *p) {
   return __atomic_load_n(p, __ATOMIC_RELAXED);
 }
